fix: download file
- new modality

+++ test.sh
@@ -16,14 +16,28 @@
 #
 # - run on intel movidius and test images folder:
 # sh nomescript.sh movidius ../test_folder
+# 
+# - run on intel movidius and donwload test images:
+# sh nomescript.sh movidius
 #
-# - run on intel movidius and single file:
-# sh nomescript.sh empty ../test.jpeg
+# - run reduce on intel movidius and single file:
+# sh nomescript.sh reduce ../test.jpeg
+#
+# - run on intel movidius and test images folder:
+# sh nomescript.sh reduced ../test_folder
+# 
+# - run on intel movidius and donwload test images:
+# sh nomescript.sh reduced
+#
+# - run test host and single image:
+# sh nomescript.sh empty ../test_images.jpg
 #
 # - run test host and folder images:
-# sh nomescript.sh empty ../test_folder
-#
-# NB empty is a fake parameter
+# sh nomescript.sh empty 
+# 
+# - run test host and and donwload test images:
+# sh nomescript.sh empty 
+# NB `empty` IS A FAKE PARAMETER
 #
 
 #---- FUNCTION DEFINITION ----------------------------------------------------#
@@ -47,7 +61,12 @@ echo  "working directory:" $PWD
 # dowload where download all file
 dest_dir=Model
 if [ ! -d "${dest_dir}" ]; then
-	echo "Make folder . . ."
+	echo "Make folder..."
+	mkdir $dest_dir
+else
+	echo "Remove previous model..."
+	rm -r ${dest_dir}
+	echo "Make folder..."
 	mkdir $dest_dir
 fi
 cd $dest_dir
@@ -56,6 +75,12 @@ cd $dest_dir
 if [ "$1" = "movidius" ]; then
 	echo "Start download of model specific for Intel Movidius"
 	model_file='1Cssy9EP8CSB6kf0tvV46o-518aLzqxYS' 
+ 	file=$(downloadfromGDrive $model_file $destdir)
+elif [ "$1" = "reduced" ]; then
+	model_file='136M-LUWRlECnnivyLCprBRC19N8mDyfx' 
+ 	file=$(downloadfromGDrive $model_file $destdir)
+elif [ "$1" = "reduced-nocompile" ]; then
+	model_file='1XfJMaMUVSVFD9ukA7oju-wFo2P2cz3uN' 
  	file=$(downloadfromGDrive $model_file $destdir)
 else
 	echo "Start download of model"
@@ -67,11 +92,25 @@ echo "decompressing file ${file}"
 zip_files="${dest_dir}/${file}"
 tar -zxvf $file
 rm $file
+find . -type f -name '._*' -delete
 
 # return parent folder and get config and model file
 cd ..
 config_file=$(find -L "${PWD}/${dest_dir}" -name \*.json | sort)
 if [ "$1" = "movidius" ]; then
+	model_file=$(find -L "${PWD}/${dest_dir}" -name \*.graph | sort)
+elif [ "$1" = "reduced-nocompile" ]; then
+	model_file=$(find -L "${PWD}/${dest_dir}" -name \*.graph | sort)
+elif [ "$1" = "reduced-keras" ]; then
+	model_file=$(find -L "${PWD}/${dest_dir}" -name \*.h5 | sort)
+elif [ "$1" = "reduced" ]; then
+	model_file=$(find -L "${PWD}/${dest_dir}" -name \*.pb | sort)
+	echo "mvNCCheck ${model_file} -in input_1 -on predictions/Softmax -s 12"	
+	mvNCCheck ${model_file} -in input_1 -on predictions/Softmax -s 12
+	echo "mvNCProfile ${model_file} -in input_1 -on predictions/Softmax -s 12"
+	mvNCProfile ${model_file} -in input_1 -on predictions/Softmax -s 12
+	echo "mvNCCompile ${model_file} -in input_1 -on predictions/Softmax -s 12 -o ${PWD}/${dest_dir}/little_vgg16.graph"
+	mvNCCompile ${model_file} -in input_1 -on predictions/Softmax -s 12 -o ${PWD}/${dest_dir}/little_vgg16.graph
 	model_file=$(find -L "${PWD}/${dest_dir}" -name \*.graph | sort)
 else
 	model_file=$(find -L "${PWD}/${dest_dir}" -name \*.pb | sort)
@@ -79,17 +118,22 @@ fi
 echo "Found: ${config_file}"
 echo "Found: ${model_file}"
 
-# download test folder contains images
-echo "Download test images..."
-remote_test_zip='16XMxZahk2OCETN-HiSVtIJkGqyNsbzJb'
-zip_test_folder=$(downloadfromGDrive $remote_test_zip)
-echo "decompressing ${zip_test_folder}..."
+# define test folder for images
 test_folder=TestImages
-mkdir -p $test_folder && tar xf $zip_test_folder -C $test_folder --strip-components=1
-rm $zip_test_folder
-
+if [ -z "$2" ] && [ ! -d "${test_folder}" ]; then
+	echo "No image/s supplied"
+	# download test folder contains images
+	echo "Download test images..."
+	remote_test_zip='16XMxZahk2OCETN-HiSVtIJkGqyNsbzJb'
+	zip_test_folder=$(downloadfromGDrive $remote_test_zip)
+	echo "decompressing ${zip_test_folder}..."
+	mkdir -p $test_folder && tar xf $zip_test_folder -C $test_folder --strip-components=1
+	rm $zip_test_folder
+	# launch script
+	echo "python3 prediction.py --configfile ${config_file} --model ${model_file} --test $2"
+	python3 prediction.py --configfile ${config_file} --model ${model_file} --test $2
+else
 # launch script
 echo "python3 prediction.py --configfile ${config_file} --model ${model_file} --test ${test_folder}"
 python3 prediction.py --configfile ${config_file} --model ${model_file} --test ${test_folder}
-
-
+fi
